soft/requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.1
opencv-python==4.13.0.90
pyserial==3.5
pyzmq==27.1.0
Werkzeug==3.1.5
